# swagger_output.json
{
  "swagger": "2.0",
  "info": {
    "title": "API Cryptocoin Exodo",
    "description": "Description",
    "version": "1.0.0"
  },
  "host": "https://criptocoin.rj.r.appspot.com",
  "basePath": "/",
...
  "schemes": [
    "http",
    "https"
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
...
    "/userPurchase/{id}": {
      "get": {
        "tags": [
          "Compras"
        ],
        "description": "Retorna as operações feita por um usuario",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "401": {
            "description": "Não autorizado"
          },
          "422": {}
        }
      }
    },
...
    "/getPurchase/{operacao}": {
      "get": {
        "tags": [
          "Compras"
        ],
        "description": "Retorna as operações feita por um usuario. Comprar, vender ou geral",
        "parameters": [
          {
            "name": "operacao",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "401": {
            "description": "Não autorizado"
          }
        }
      }
    },
...
    "/purchase": {
      "post": {
        "tags": [
          "Compras"
        ],
        "description": "Registra uma operação. comprar ou vender",
        "parameters": [
          {
            "name": "obj",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "operacao": {
                  "type": "string",
                  "example": "any"
                },
                "usuario": {
                  "type": "string",
                  "example": "any"
                },
                "moeda": {
                  "type": "string",
                  "example": "any"
                },
                "quantidade": {
                  "type": "string",
                  "example": "any"
                },
                "valor": {
                  "type": "string",
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "401": {
            "description": "Não autorizado"
          }
        }
      }
    },
    "/payment/comprar": {
      "post": {
        "tags": [
          "Compras"
        ],
        "description": "Confirma uma compra",
        "parameters": [
          {
            "name": "obj",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "_id": {
                  "type": "string",
                  "example": "any"
                },
                "usuarioEmissor": {
                  "type": "string",
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "401": {
            "description": "Não autorizado"
          }
        }
      }
    },
    "/payment/vender": {
      "post": {
        "tags": [
          "Compras"
        ],
        "description": "Confirma uma venda",
        "parameters": [
          {
            "name": "obj",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "_id": {
                  "type": "string",
                  "example": "any"
                },
                "usuarioReceptor": {
                  "type": "string",
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "401": {
            "description": "Não autorizado"
          }
        }
      }
    },
    "/payment/recusar": {
      "post": {
        "tags": [
          "Compras"
        ],
        "description": "Rejeita uma compra ou venda",
        "parameters": [
          {
            "name": "obj",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "_id": {
                  "type": "string",
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "401": {
            "description": "Não autorizado"
          }
        }
      }
    },
    "/sacar/all": {
      "get": {
        "tags": [
          "Saque"
        ],
        "description": "Retorna todos os saques",
        "parameters": [
          {
            "name": "obj",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "taxa": {
                  "type": "string",
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "401": {
            "description": "Não autorizado"
          },
          "422": {}
        }
      }
    },
    "/sacar": {
      "get": {
        "tags": [
          "Saque"
        ],
        "description": "Retorna os saques do usuario",
        "parameters": [],
        "responses": {
          "401": {
            "description": "Não autorizado"
          },
          "422": {}
        }
      },
      "post": {
        "tags": [
          "Saque"
        ],
        "description": "Efetua o saque",
        "parameters": [
          {
            "name": "obj",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "moedaNome": {
                  "type": "string",
                  "example": "any"
                },
                "valorSaque": {
                  "type": "string",
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "401": {
            "description": "Não autorizado"
          },
          "422": {}
        }
      }
    },
    "/sacar/{id}": {
      "get": {
        "tags": [
          "Saque"
        ],
        "description": "Retorna o saque pelo Id do saque",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "401": {
            "description": "Não autorizado"
          },
          "422": {}
        }
      }
    },
...
  },
...
}